Make data standardizer

Create a data standardizer function using the relevant statistics in a input data frame.

make_standardizer(x)

Arguments

x
a base::data.frame() or similar object

Value

The returned object is a function that will standarize all non-binary variables using the means and SDs for relevant columns in the input x data frame.

The point of doing it this way is that we need to record the relevant training data means and SDs so we can standardize test data at the prediction stage as well.

One option would be to record the trainin data means and SDs instead, but we will in any case use that to call another function that does the standardization, so we might as well do it now.

Examples

library("dplyr")
library("stats")
data("states")

# assume that the input consists entirely of features, and no NAs
train_x <- states %>%
  filter(year < 2010) %>%
  select(-starts_with("dv_"), -gwcode, -year) %>%
  filter(complete.cases(.))

test_x <- states %>%
  filter(year > 2009) %>%
  select(-starts_with("dv_"), -gwcode, -year) %>%
  filter(complete.cases(.))

standardize <- make_standardizer(train_x)
prepped_train_x <- standardize(train_x)

do.call(rbind,
        lapply(prepped_train_x, function(x) {
          data.frame(mean = mean(x), SD = sd(x))
        }))

#>                                     mean        SD
#> v2x_veracc_osp             -1.195332e-16 1.0000000
#> v2x_veracc_osp_diff_y2y     1.983831e-18 1.0000000
#> v2xcs_ccsi                  6.511823e-17 1.0000000
#> v2xcs_ccsi_diff_y2y         9.588867e-18 1.0000000
#> v2xcl_rol                  -3.556932e-17 1.0000000
#> v2xcl_rol_diff_y2y          7.411389e-19 1.0000000
#> v2x_freexp_altinf          -1.555001e-16 1.0000000
#> v2x_freexp_altinf_diff_y2y  3.667081e-18 1.0000000
#> v2x_horacc_osp              1.364462e-16 1.0000000
#> v2x_horacc_osp_diff_y2y     1.465993e-18 1.0000000
#> v2x_pubcorr                -5.163898e-18 1.0000000
#> v2x_pubcorr_diff_y2y       -3.353040e-18 1.0000000
#> binary_x                    2.017315e-01 0.4013255

# the training data means and SDs are used for standardization
prepped_test_x <- standardize(test_x)
do.call(rbind,
        lapply(prepped_test_x, function(x) {
          data.frame(mean = mean(x), SD = sd(x))
}))

#>                                   mean        SD
#> v2x_veracc_osp              0.45301148 0.7447701
#> v2x_veracc_osp_diff_y2y    -0.05491269 0.5904957
#> v2xcs_ccsi                  0.43033773 0.8324992
#> v2xcs_ccsi_diff_y2y        -0.16444419 0.9515943
#> v2xcl_rol                   0.34756717 0.8715744
#> v2xcl_rol_diff_y2y         -0.14202215 0.8872663
#> v2x_freexp_altinf           0.42522854 0.8115048
#> v2x_freexp_altinf_diff_y2y -0.16133401 0.8358461
#> v2x_horacc_osp              0.36538211 0.8965973
#> v2x_horacc_osp_diff_y2y    -0.06061840 0.8259145
#> v2x_pubcorr                 0.05373219 1.0049355
#> v2x_pubcorr_diff_y2y       -0.12548883 1.1674437
#> binary_x                    0.20013167 0.4002305